ServiceWrapperBuilderTest introduced
